Moving test case to the same file
